requirements.txt
streamlit
crewai
crewai_tools
openai
python-dotenv
serper
